requirements.txt
flask
flask-cors
stripe
gunicorn
werkzeug
psycopg2-binary